handle next(err) even if middleware returns undefined
Previously next(err) worked only if a the rejected promise created by
next(err) was returned from the middleware. This change makes the
expected behaviour closer to Express’, but disallow the user to handle
errors with next().catch() in middleware before the one that produced
the error. So it’s a breaking change.